External data scalar converted from single to double
git-svn-id: https://sources.ecopath.org/svn/Ecopath/trunk/Ecopath6@7818 a725721c-0ed6-864e-a6a8-02c2a1988c9d

+++ Sources/EwECore/SpatialTemporal/Adapters/cRelPPDataAdapter.vb
@@ -20,7 +20,7 @@ Namespace SpatialData
 
 #Region " Private vars "
 
-        Private m_sPreservedScale As Single = cCore.NULL_VALUE
+        Private m_sPreservedScale As Double = cCore.NULL_VALUE
         Private m_spaceData As cEcospaceDataStructures
 
 #End Region ' Private vars
